refactor(llm): extract the hosted proxy error mapper into util/hosted-error so the STT and TTS providers can stop surfacing raw proxy JSON, one definition for every hosted surface
