+++ src/llm/usejarvis.ts
@@ -1,5 +1,5 @@
 import { OpenAIProvider } from './openai.ts';
-import { redactSecrets } from '../util/redact.ts';
+import { hostedProxyError } from '../util/hosted-error.ts';
 
 /**
  * Hosted "Usejarvis AI" provider: the platform's OpenAI-compatible LLM proxy.
@@ -81,47 +81,12 @@ export class UsejarvisAIProvider extends OpenAIProvider {
     }
   }
 
-  /** Map proxy errors to actionable copy. The `(status)` marker is kept so
-   * classifyErrorString still steers retries (429/503 retry; 400s do not).
-   *
-   * Branch ORDER matters: LiteLLM denies an out-of-plan model with a 401
-   * "not allowed to access model", so the model check MUST run before the
-   * generic auth branch — otherwise a paying user who picks a model outside
-   * their plan is told their subscription is inactive, the worst false
-   * message this feature can produce.
-   *
-   * The proxy's own body never rides along in the returned copy: it is not a
-   * log-only channel — this Error's message becomes the chat bubble — and an
-   * upstream body carries the hosted hostname that the settings surface and
-   * the catalog route both deliberately withhold. Operators get it via
-   * console.warn instead. */
+  /** Shared with the hosted STT/TTS providers — see util/hosted-error.ts for
+   * the redaction, branch-order and no-body-in-copy rules. The "<label> API"
+   * form keeps the exact `Usejarvis AI API error (NNN)` prefix that
+   * classifyErrorString and rewriteText both parse. */
   private friendly(status: number, detail: string): Error {
-    // Redact FIRST: proxy auth bodies can echo the bearer we presented, and
-    // this per-account key is deliberately hidden from every other surface
-    // (settings, catalog route, provider test).
-    const safe = redactSecrets(detail);
-    const lower = safe.toLowerCase();
-    if (safe) console.warn(`[usejarvis] proxy error (${status}): ${safe.slice(0, 200)}`);
-    if (lower.includes('budget') && (lower.includes('exceed') || lower.includes('over'))) {
-      return new Error(
-        `${this.errorLabel} API error (${status}): your included AI usage is used up ` +
-          `${describeBudgetWindow(safe)}. It resumes automatically - the usage meter shows when.`,
-      );
-    }
-    if (lower.includes('model') && (lower.includes('not allowed') || lower.includes('invalid model'))) {
-      return new Error(
-        `${this.errorLabel} API error (${status}): that model is not included in your plan.`,
-      );
-    }
-    if (status === 401 || status === 403) {
-      return new Error(
-        `${this.errorLabel} API error (${status}): Usejarvis AI is not active on this account - ` +
-          'an active plan is required.',
-      );
-    }
-    // Truncated: an unbounded body is how a CDN error page (hostname
-    // included) reaches a chat bubble.
-    return new Error(`${this.errorLabel} API error (${status})${safe ? `: ${safe.slice(0, 120)}` : ''}`);
+    return hostedProxyError(`${this.errorLabel} API`, status, detail);
   }
 
   /** Text-shaped variant of `rewrite` for stream error EVENTS (the base class
@@ -142,25 +107,3 @@ export class UsejarvisAIProvider extends OpenAIProvider {
   }
 }
 
-/**
- * Turn a proxy budget body into the window phrase the friendly copy promises.
- * LiteLLM reports `budget_duration` and `budget_reset_at` on an exhausted
- * key, and the reset lands on a FIXED clock boundary (a 6h window minted
- * mid-morning resets at 12:00:00+00:00), so "resumes at 12:00 UTC" is exact
- * rather than approximate. Degrades to the generic phrase when the proxy
- * omits the fields — never guesses a time it was not told.
- *
- * Only the duration and timestamp are lifted out; the rest of the body stays
- * out of user-facing copy (it can carry the hosted hostname).
- */
-function describeBudgetWindow(body: string): string {
-  const duration = body.match(/budget_duration["'\s:=]+([0-9]+[a-z]+)/i)?.[1];
-  const resetAt = body.match(/budget_reset_at["'\s:=]+([0-9T:.+\-]{10,40})/i)?.[1];
-  const window = duration ? `for this ${duration} window` : 'for this window';
-  if (!resetAt) return window;
-  const parsed = new Date(resetAt);
-  if (Number.isNaN(parsed.getTime())) return window;
-  const hh = String(parsed.getUTCHours()).padStart(2, '0');
-  const mm = String(parsed.getUTCMinutes()).padStart(2, '0');
-  return `${window} (resumes ${hh}:${mm} UTC)`;
-}
